Rewrite key computation and representation
Base key computation on the SHA-256 of the skeleton key instead of using
the skeleton key directly. This make it possible to save the hash
locally in encrypted form later on without compromising the skeleton
key.

Start key generation with the (hash of the) skeleton key so that the
hashing algorithm cannot be primed with know service and user names when
brute forcing.

Instead of producing single key produce a chain of keys. When an
individual key has been compromized (database cracked) it can be
replaced with the next one in the list.

Truncate the produced keys to 80 bits and encode them in a format based
on Cockford's base 32 encoding. This is a balanced compromize between
readability, writability, some degree of memorizability, and password
restrictions of different services (e.g. Skype limits passwords to 20
characters, WTF).